pyproject.toml
[project]
name = "cthulhu-news"
version = "0.1.0"
description = "Cthulhu-inspired news"
authors = [
    { name = "Mysterious Ben", email = "datascience@tuta.io" }
]
readme = "README.md"
requires-python = ">=3.12.4,<3.13"
dependencies = [
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "envparse>=0.2.0",
    "logutil[loguru]>=0.1.11",
    "apscheduler>=3.10.4",
    "httpx>=0.25.2",
    "pymongo>=4.6.1",
    "openai>=1.3.8",
    # "qdrant-client[fastembed]>=1.6.9",  # uncomment if needed
    "pillow>=10.1.0",
    "python-multipart>=0.0.6",
    "spacy>=3.7.5",
    "lxml[html-clean]>=5.2.2",
    "tenacity>=9.0.0",
    "litellm>=1.72.6.post1",
    "anthropic>=0.54.0",
    "prefect>=3.4.6",
]

[project.optional-dependencies]
db = [
    "newspaper3k>=0.2.8"
]
web = [
    "cachetools>=5.5.0",
    "fastapi>=0.104.1",
    "jinja2>=3.1.2",
    "psycopg[pool]>=3.2.3",
    "uvicorn[standard]>=0.24.0.post1",
    "pgvector>=0.3.6"
]
dev = [
    "pylint>=3.0.2",
    "pytest>=7.4.3",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
    "psycopg[binary]>=3.2.3",
    "ruff>=0.12.0"
]
jupyter = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.1",
    "pandas>=2.2.2"
]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.ruff]
line-length = 99
target-version = "py310"
exclude = [
    ".eggs",
    ".git",
    ".github",
    ".mypy_cache",
    ".tox",
    ".venv",
    "notebooks",
    "resources",
    ".ipynb_checkpoints",
    "__pycache__",
    "_build",
    "build",
    "dist",
    ".venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
markers = [
    "integration: integration testing (tests require valid credentials and can be slow)",
]
testpaths = [
    "tests",
]

[tool.mypy]
enable_incomplete_feature = ["InlineTypedDict"]
